add a task_id in the log
franckly I'm not overly convinced by this, but I thought it would be
easier to debug if we can grep on the datagouv id in the log.

the main problem is that `subprocess` does not seems to use the logger,
so the gtfs2netexfr logs does not have the task_id.

Also in order to have an optional field, I switched to json log, to the
log are now:

```
{"asctime": "2020-05-04 08:23:24,757", "levelname": "DEBUG", "message":
"Got a netex repooo GTFS-20190612-171806.zip.netex.zip", "task_id":
"5d04357406e3e71839d1585a"}
```

feel free to disregard if you think it's useless